+++ CMS-iOS/Support/SpotlightIndex.swift
@@ -3,7 +3,7 @@
 //  CMS-iOS
 //
 //  Created by Aryan Chaubal on 7/7/20.
-//  Copyright © 2020 Hridik Punukollu. All rights reserved.
+//  Copyright © 2020 Crux BPHC. All rights reserved.
 //
 
 import CoreSpotlight