+++ CMS-iOS/View Controllers/LoginViewController.swift
@@ -3,7 +3,7 @@
 //  CMS-iOS
 //
 //  Created by Hridik Punukollu on 09/08/19.
-//  Copyright © 2019 Hridik Punukollu. All rights reserved.
+//  Copyright © 2019 Crux BPHC. All rights reserved.
 //
 
 import UIKit
@@ -200,6 +200,10 @@ class LoginViewController: UIViewController {
         let decodedUserIDSub = (decodedString[start..<end])
         let decodedUserID = String(decodedUserIDSub)
         print("decoded = \(decodedUserID)")
+        let privateTokenStart = decodedString.index(decodedString.startIndex, offsetBy: 68)
+        let privateTokenEnd = decodedString.endIndex
+        let privateToken = String(decodedString[privateTokenStart..<privateTokenEnd]).replacingOccurrences(of: "::", with: "")
+        KeychainWrapper.standard.set(privateToken, forKey: "privateToken")
         logIn (password: decodedUserID, loggedin: false) {
             print("Password Retrieved. Logging in.")
             self.view.isUserInteractionEnabled = true